Add main classroom HTTP APIs
