+++ routes/api/apiPermissionCheck.js
@@ -1,4 +1,4 @@
-const { classInformation, getClassIDFromCode } = require("../../modules/class")
+const { classInformation, getClassIDFromCode } = require("../../modules/class/classroom")
 const { getUser } = require("../../modules/user")
 const { logger } = require("../../modules/logger")
