+++ routes/api/class/permissions.js
@@ -1,4 +1,4 @@
-const { classInformation } = require("../../../modules/class")
+const { classInformation } = require("../../../modules/class/classroom")
 const { logger } = require("../../../modules/logger")
 
 module.exports = {
