+++ routes/api/class/polls/clearPoll.js
