+++ routes/api/class/polls/createPoll.js
@@ -1,14 +1,14 @@
 const { createSocketFromHttp } = require("../../../../modules/webServer");
 const { createPoll } = require("../../../../modules/polls");
 const { logger } = require("../../../../modules/logger");
-const {httpPermCheck} = require("../../../middleware/permissionCheck");
+const { httpPermCheck } = require("../../../middleware/permissionCheck");
 
 module.exports = {
     run(router) {
         // Creates a poll from the data provided
         router.post('/class/:id/polls/create', async (req, res) => {
             try {
-                const hasPerms = await httpPermCheck("clearPoll", req, res);
+                const hasPerms = await httpPermCheck("createPoll", req, res);
                 if (!hasPerms) {
                     return;
                 }
