+++ routes/api/class/polls/endPoll.js
@@ -1,14 +1,14 @@
 const { createSocketFromHttp } = require("../../../../modules/webServer");
 const { endPoll } = require("../../../../modules/polls");
 const { logger } = require("../../../../modules/logger");
-const {httpPermCheck} = require("../../../middleware/permissionCheck");
+const { httpPermCheck } = require("../../../middleware/permissionCheck");
 
 module.exports = {
     run(router) {
         // Ends the current poll for the class
         router.post('/class/:id/polls/end', async (req, res) => {
             try {
-                const hasPerms = await httpPermCheck("clearPoll", req, res);
+                const hasPerms = await httpPermCheck("endPoll", req, res);
                 if (!hasPerms) {
                     return;
                 }
