+++ routes/apikey.js
@@ -1,5 +1,5 @@
 const { isLoggedIn } = require('../modules/authentication');
-const { classInformation } = require('../modules/class');
+const { classInformation } = require('../modules/class/classroom');
 const { logNumbers } = require('../modules/config');
 const { logger } = require('../modules/logger');
