+++ routes/tests/student.spec.js
@@ -1,7 +1,7 @@
 const studentRoute = require('../student');
 const request = require('supertest');
 const { createExpressServer } = require('../../modules/tests/tests');
-const { classInformation } = require('../../modules/class');
+const { classInformation } = require('../../modules/class/classroom');
 const { STUDENT_PERMISSIONS, GUEST_PERMISSIONS } = require('../../modules/permissions');
 
 describe('Student Route', () => {
