+++ sockets/break.js
@@ -1,4 +1,4 @@
-const { classInformation } = require("../modules/class")
+const { classInformation } = require("../modules/class/classroom")
 const { logger } = require("../modules/logger")
 const { advancedEmitToClass } = require("../modules/socketUpdates")
 const { io } = require("../modules/webServer")
