+++ sockets/digipogs.js
@@ -1,5 +1,5 @@
 const { transferDigipogs } = require('../modules/digipogs');
-const { classInformation } = require('../modules/class');
+const { classInformation } = require('../modules/class/classroom');
 const { database } = require('../modules/database');
 const {logger} = require("../modules/logger");
