+++ sockets/timer.js
@@ -1,4 +1,4 @@
-const { classInformation } = require("../modules/class")
+const { classInformation } = require("../modules/class/classroom")
 const { logger } = require("../modules/logger")
 const { CLASS_SOCKET_PERMISSIONS } = require("../modules/permissions")
 const { advancedEmitToClass, runningTimers } = require("../modules/socketUpdates")
