+++ sockets/updates.js
@@ -1,4 +1,4 @@
-const { classInformation } = require("../modules/class")
+const { classInformation } = require("../modules/class/classroom")
 const { logger } = require("../modules/logger")
 const { managerUpdate } = require("../modules/socketUpdates")
